move all of main tests into cli
